BUG: Remove VNL views
VNL vectors and matrices only have constructors that copy the input data. It is
therefore not possible to create a view on NumPy arrays with VNL vectors and
matrices. These functions were not working and are therefore removed.
It is still possible to create views on VNL vectors and matrices with NumPy
arrays.